Added support for model specific hooks (#166)
refs https://github.com/TryGhost/Arch/issues/73

Up until now we have used hooks at a global level, where we intercept the plugin
to add logic for particular table names. But we now have a usecase where we want
to add certain logic based on properties of the model. This extends the plugin
to be able to read hooks from the model. It is backwards compatible, so will
continue to work as-is for existing setups.

Note that we do not use the _.has helper as this only read properties directly
from the object, and not from the prototype chain, which will not work for
models which would have these hooks on the prototype.